Weedu/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.3
django-cors-headers==4.6.0
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
